requirements.txt
asgiref==3.2.3
certifi==2019.11.28
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.4
pytz==2019.3
sentry-sdk==0.14.2
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.0.1
